requirement.txt
streamlit>=1.28.0
praw>=7.7.1
python-dotenv>=1.0.0
openai>=1.3.0
chromadb>=0.4.15
sentence-transformers>=2.2.2